requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pillow==10.4.0
opencv-python==4.10.0.84
transformers==4.46.0
torch==2.5.1
accelerate==1.0.1
openpyxl==3.1.5
reportlab==4.2.5
numpy==2.1.2
